+++ international/en.php
@@ -641,7 +641,9 @@ $languages['en'] = array (
 	"label_communityradio" => "Community Radio Browser",
 	"label_volume" => "Volume",
 	"config_playlistswipe" => "Swipe to remove tracks from Current Playlist",
-	"button_debuginfo" => "Debug Information"
+	"button_debuginfo" => "Debug Information",
+	"config_podcast_defaults" => "Default Options For New Podcasts",
+	"config_marknewasunlistened" => "Mark New Episodes as Unlistened Before Refresh"
 );
 
 ?>